moving core/__init__.py to core/test_executor.py
